+++ sqlprojection/adaptor.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/dogmatiq/dogma"
 	"github.com/dogmatiq/projectionkit/internal/identity"
-	"github.com/dogmatiq/projectionkit/internal/unboundhandler"
 	"github.com/dogmatiq/projectionkit/resource"
 )
 
@@ -31,10 +30,6 @@ func New(
 	d Driver,
 	h MessageHandler,
 ) dogma.ProjectionMessageHandler {
-	if db == nil {
-		return unboundhandler.New(h)
-	}
-
 	return &adaptor{
 		db:      db,
 		handler: h,